% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comtrade_database_downloader.R
\docType{data}
\name{comtrade_names}
\alias{comtrade_names}
\title{UN to ISO 3166 Alpha-3 country code correspondence table}
\format{
A list
}
\source{
\url{https://unstats.un.org/wiki/display/comtrade/Country+codes+in+ISO+3166}
}
\usage{
comtrade_names
}
\description{
UN to ISO 3166 Alpha-3 country code correspondence table
}
\keyword{datasets}
