% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comtrade_database_downloader.R
\name{get_comtrade_file}
\alias{get_comtrade_file}
\title{Download Comtrade Database Period}
\usage{
get_comtrade_file(freq, year, month = NULL, token)
}
\description{
Part of the \pkg{ComtradeDatabaseDownloader} package. \cr
A support function for the \code{\link{get_comtrade}} function.
}
