% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comtrade_database_downloader.R
\name{wrangle_comtrade_data}
\alias{wrangle_comtrade_data}
\title{Wrangle Comtrade Database Period}
\usage{
wrangle_comtrade_data(df, freq)
}
\description{
Part of the \pkg{ComtradeDatabaseDownloader} package. \cr
A support function for the \code{\link{get_comtrade}} function.
}
